move types to another file to make API generation work
